edit create latest

+++ scripts/create_latest_linux_aarch64_player.sh
@@ -2,7 +2,10 @@
 
 LINK_PATH="/home/niko/downloads/ci-builds"
 
-rm $LINK_PATH/latest_linux_aarch64_player.AppImage
+if [ -f $LINK_PATH/latest_linux_aarch64_player.AppImage ]; then
+  rm $LINK_PATH/latest_linux_aarch64_player.AppImage
+fi
+
 TARGET=`ls -rt $LINK_PATH/garlic-player_linux/garlic-player-linux-aarch64*AppImage | tail -n1`
 ln -s $TARGET $LINK_PATH/latest_linux_aarch64_player.AppImage
