+++ scripts/create_latest_macosx_player.sh
@@ -2,6 +2,8 @@
 
 LINK_PATH="/home/niko/downloads/ci-builds"
 
-rm $LINK_PATH/latest_macosx_player.dmg
+if [ -f $LINK_PATH/latest_macosx_player.dmg ]; then
+  rm $LINK_PATH/latest_macosx_player.dmg
+fi
 
 ln -s $LINK_PATH/garlic-player_macosx/`ls -rt $LINK_PATH/garlic-player_macosx/ | tail -n1` $LINK_PATH/latest_macosx_player.dmg
