+++ scripts/create_latest_rpi64_player_image.sh
@@ -2,5 +2,8 @@
 
 LINK_PATH="/home/niko/downloads/ci-builds"
 
-rm $LINK_PATH/latest_raspberrypi64.img.xz
+if [ -f $LINK_PATH/latest_raspberrypi64.img.xz ]; then
+  rm $LINK_PATH/latest_raspberrypi64.img.xz
+fi
+
 ln -s $LINK_PATH/garlic-player_rpi64/`ls -rt $LINK_PATH/garlic-player_rpi64/ | tail -n1` $LINK_PATH/latest_raspberrypi64.img.xz
